SERIOUS problems with the fretting system. It seems everything is working, but there are too many things to be fixed D= Have to take a look at the code and undestand why so many strange things are going on. BUT... Now we have the following hierachy: Player (Track (Fretting (EventReceiver)), Fretting (EventReceiver)) --> both the 2 Fretting and EventReceiver are the same (just pointers). The only asynchronous work done on the EventReceiver class is the event attribution. Other things are done on Player->update. (too much thing to be fixed yet, as said before D=)

+++ trunk/Track.h
@@ -44,7 +44,7 @@ class Track
 		Fretting		*fretting;
 		
 		// used for update()
-		int				tolerance;		// to warn when the balls passed their time without being pressed
+		double			tolerance;		// to warn when the balls passed their time without being pressed
 		double			chordCreationTime;
 		bool			nonPressedChord;	// indicates if the player didn't press a chord
 		int				notesOnChord;		// how many notes did we have on the chord??
